Tales improvement for vegetables
I have the feeling that the tales need more words in order to work and sound properly.
This commit  shall work as an example how we could improve it.
I splitted up the Vegetables.txt into
->Vegetables_Singular.txt
->Vegetables_Singular_Feminine.txt
->Vegetables_Singular_Masculine.txt
->Vegetables_Plural.txt

I replaced the content of the translated Vegetables.txt with the original English words (just for the information; the file itself will not be used anymore)

In RulePacks_Global.xml I introduced the words in the files to the game.
Results:
[VegetableSingular] -> (for cases where the gender is not important)
[VegetableSingularFem] -> (for feminine vegetables such as Karotte)
[VegetableSingularMas] -> (for masculine vegetables such as Kürbis)
[VegetablePlural] -> (for plural forms of every vegetable; the gender is not important there)

Then I replaced every instance of [Vegetable] with the fitting text.  Now every mentioned vegetable in the tales/game should have the right grammatical form. (Maybe we still would need something for the accusative later; but for now it should suffice)

Before: "[desc_pawn] [Character], der ein Linse isst"
("ein" was used for every vegetable, even the feminine ones)

After: "[desc_pawn] [Character], der eine Linse isst"

If you like the changes and if they work in the game we can do this to other nouns as well, such as animals, jobs etc...