Changed RulePacks_Namers_Factions
Explanation for changes in GroupNames.txt:
"Haufen" was changed to "Legion" because it was the only non-feminine word in the list. This would have unnecessarily complicated the generation of correct sounding names.

Explanation for changes in Concepts_Badass.txt:
"Zwist" was changed to "Zwistigkeit" because it was the only non-feminine word in the list. This would have unnecessarily complicated the generation of correct sounding names.

The name generation should give better results now.