Removed redundant "Nouns" folder (moved its content to old "Nouns" folder)
I noticed that the changes in the commits
8450a45
and
6102ce8
were done on  a redundant (newly created) "RimWorld-de\Nouns" folder which has no excuse to be there.
So I moved its contents to "RimWorld-de\Strings\Words\Nouns" and deleted the redundant folder.

Everything works as it did before. It's only less complicated now.