Fixed [damaged_inf] in Combat Logs
In English the infinite verb ca be used as a noun. In German it's not so easy:
((to) scratch = (zu) kratzen; scratch = Kratzer)
Because of that I fixed some instances of [damaged_inf] where it was used as a noun.

Todo: There are still a lot of other things to fix in the combat logs.