Add GitHub action to auto-run update-wordinfo
This prevents the need to install certain tools to run the script, depending on the OS. New words in new_words.txt must still be manually sorted into Male.txt, Female.txt, Neuter.txt or Plural.txt. So far I have not found a free API that reliably determines the gender of a German word.